requirements.txt
python-dotenv
pandas 
numpy
pymongo[srv]==3.6
pymongo
certifi
scikit-learn
-e . 